feat: Organize event namespaces and update consumers for better structure and clarity
