+++ services/orders/Orders.Application/Events/Integration/Inventory/InventoryReservedIntegrationEvent.cs
@@ -1,6 +1,6 @@
 using Shared.Domain.Events;
 
-namespace Orders.Application.Events;
+namespace Orders.Application.Events.Integration.Inventory;
 
 public record InventoryReservedIntegrationEvent(
     Guid OrderId,
