+++ services/orders/Orders.Application/Events/Integration/Order/OrderStatusChangedIntegrationEvent.cs
@@ -1,6 +1,6 @@
 using Shared.Domain.Events;
 
-namespace Orders.Application.Events;
+namespace Orders.Application.Events.Integration.Order;
 
 public record OrderStatusChangedIntegrationEvent(
     Guid OrderId,
