+++ services/orders/Orders.Application/Events/Integration/Payment/PaymentRefundFailedIntegrationEvent.cs
@@ -1,6 +1,6 @@
 using Shared.Domain.Events;
 
-namespace Orders.Application.Events;
+namespace Orders.Application.Events.Integration.Payment;
 
 public record PaymentRefundFailedIntegrationEvent(
     Guid OrderId,
