+++ services/orders/Orders.Application/Events/Integration/Shipment/ShipmentFailedIntegrationEvent.cs
@@ -1,6 +1,6 @@
 using Shared.Domain.Events;
 
-namespace Orders.Application.Events;
+namespace Orders.Application.Events.Integration.Shipment;
 
 public record ShipmentFailedIntegrationEvent(
     Guid OrderId,
