+++ services/orders/Orders.Application/Messaging/Consumers/PaymentProcessedCommandConsumer.cs
@@ -1,11 +1,12 @@
 using MassTransit;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Orders.Application.Messaging.Contracts;
 // using Payment.Api.Events; // Cross-service event
 
 namespace Orders.Infrastructure.Messaging.Consumers
 {
-    public class PaymentProcessedConsumer : IConsumer<PaymentProcessedIntegrationEvent>
+    public class PaymentProcessedConsumer : IConsumer<ProcessPaymentCommand>
     {
         private readonly ILogger<PaymentProcessedConsumer> _logger;
         // Inject your application services here
@@ -15,7 +16,7 @@ namespace Orders.Infrastructure.Messaging.Consumers
             _logger = logger;
         }
 
-        public async Task Consume(ConsumeContext<PaymentProcessedIntegrationEvent> context)
+        public async Task Consume(ConsumeContext<ProcessPaymentCommand> context)
         {
             _logger.LogInformation("Processing payment completed for Order: {OrderId}",
                 context.Message.OrderId);
