+++ services/orders/Orders.Application/Messaging/Contracts/CreateShipmentCommand.cs
@@ -1,6 +1,6 @@
 using Shared.Domain.Events;
 
-namespace Orders.Application.Events;
+namespace Orders.Application.Messaging.Contracts;
 
 public record CreateShipmentCommand(
     Guid OrderId,
